consolidate pages/graphPF into pages/graph
