+++ frontend/src/pages/Graph/GraphLegendDataPF.ts
