+++ frontend/src/pages/Graph/GraphLegendPF.tsx
