+++ frontend/src/pages/Graph/MiniGraphCardPF.tsx
