+++ frontend/src/pages/Graph/NodeDecorator.tsx
