+++ frontend/src/pages/Graph/TrafficAnimation/AnimationEdge.tsx
