+++ frontend/src/pages/Graph/TrafficAnimation/AnimationTimerConfig.ts
