+++ frontend/src/pages/Graph/TrafficAnimation/TrafficPointRendererPF.tsx
