+++ frontend/src/pages/Graph/components/stylesComponentFactory.tsx
