+++ frontend/src/pages/Graph/elements/diamond.tsx
