+++ frontend/src/pages/Graph/elements/elementFactory.ts
