+++ frontend/src/pages/Graph/elements/extendedBaseEdge.ts
