+++ frontend/src/pages/Graph/elements/triangle.tsx
