+++ frontend/src/pages/Graph/styles/styleGroup.tsx
